requirements.txt
numpy
opencv-python
scikit-learn
streamlit
Pillow
numpy
opencv-python
scikit-learn
streamlit
Pillow
